Moving all the things back to the powerfilter dir until we are ready to rename all the things.
