+++ repose-aggregator/core/repose-core/src/test/scala/org/openrepose/powerfilter/ReposeFilterChainTest.scala
@@ -17,7 +17,7 @@
  * limitations under the License.
  * =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
  */
-package org.openrepose.core
+package org.openrepose.powerfilter
 
 import java.util.concurrent.TimeUnit
 
@@ -35,7 +35,7 @@ import org.mockito.ArgumentCaptor
 import org.mockito.Matchers.{any, anyLong, same, eq => meq}
 import org.mockito.Mockito.{verify, when}
 import org.openrepose.commons.utils.servlet.http.HttpServletRequestWrapper
-import org.openrepose.core.ReposeFilterChain.{FilterContext, TracingHeader}
+import org.openrepose.powerfilter.ReposeFilterChain.{FilterContext, TracingHeader}
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.mock.MockitoSugar
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
@@ -45,7 +45,7 @@ import scala.collection.JavaConverters._
 
 @RunWith(classOf[JUnitRunner])
 class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with BeforeAndAfterEach {
-  var mockFilter: Filter =_
+  var mockFilter: Filter = _
   var mockRequest: MockHttpServletRequest = _
   var mockResponse: MockHttpServletResponse = _
   var originalChain: FilterChain = _
@@ -66,11 +66,12 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
 
   describe("doFilter") {
     it("should pass an empty chain to the next filter when there is only one remaining") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -82,12 +83,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should pass the tail of the chain onto the next filter") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
-                                                   FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -99,12 +102,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should skip filters that don't pass the check") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => false),
-                                                   FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => false),
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -116,11 +121,12 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should go to the original filter chain if it's empty") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -130,12 +136,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should skip the whole chain if the bypass url is hit") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => false),
-                                                   FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              Option(".*/bar"),
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => false),
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
+        originalChain,
+        Option(".*/bar"),
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -145,12 +153,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should go into the the chain when the bypass url isn't hit") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
-                                                   FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              Option(".*/butts"),
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
+        originalChain,
+        Option(".*/butts"),
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -165,11 +175,12 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
       listAppender.clear()
       Configurator.setLevel(ReposeFilterChain.IntrafilterLog.getName, Level.TRACE)
 
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -184,11 +195,12 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
       listAppender.clear()
       Configurator.setLevel(ReposeFilterChain.IntrafilterLog.getName, Level.TRACE)
 
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -198,12 +210,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should report metrics around a filter") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
-                                                   FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -211,11 +225,12 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should report metrics around the origin service") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
 
@@ -223,38 +238,42 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should add the metrics to the response when requested around a filter") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
-                                                   FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
       mockRequest.addHeader(TracingHeader, "true")
 
       filterChain.doFilter(mockRequest, mockResponse)
 
-      mockResponse.getHeaderNames should contain ("X-foo-Time")
+      mockResponse.getHeaderNames should contain("X-foo-Time")
     }
 
     it("should add the metrics to the response when requested around the origin service") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
       mockRequest.addHeader(TracingHeader, "true")
 
       filterChain.doFilter(mockRequest, mockResponse)
 
-      mockResponse.getHeaderNames should contain ("X-origin-Time")
+      mockResponse.getHeaderNames should contain("X-origin-Time")
     }
 
     it("should not add the metrics to the response when requested and the response is already committed") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
       mockRequest.addHeader(TracingHeader, "true")
       mockResponse.setCommitted(true)
 
@@ -264,24 +283,26 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should return a 500 when an exception occurs") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
       when(originalChain.doFilter(any[HttpServletRequest], any[HttpServletResponse])).thenThrow(new RuntimeException("test exception"))
 
       filterChain.doFilter(mockRequest, mockResponse)
 
-      mockResponse.getStatus should be (SC_INTERNAL_SERVER_ERROR)
+      mockResponse.getStatus should be(SC_INTERNAL_SERVER_ERROR)
     }
 
     it("should log if there was a an exception and the response was already committed") {
-      val filterChain = new ReposeFilterChain(List.empty,
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List.empty,
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
       when(originalChain.doFilter(any[HttpServletRequest], any[HttpServletResponse])).thenThrow(new RuntimeException("test exception"))
       mockResponse.setCommitted(true)
       val loggerContext = LogManager.getContext(false).asInstanceOf[LoggerContext]
@@ -295,12 +316,14 @@ class ReposeFilterChainTest extends FunSpec with Matchers with MockitoSugar with
     }
 
     it("should create a span with the filter name") {
-      val filterChain = new ReposeFilterChain(List(FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
-                                                   FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
-                                              originalChain,
-                                              None,
-                                              metricsRegistry,
-                                              tracer)
+      val filterChain = new ReposeFilterChain(
+        List(
+          FilterContext(mockFilter, "foo", (request: HttpServletRequest) => true),
+          FilterContext(mock[Filter], "bar", (request: HttpServletRequest) => true)),
+        originalChain,
+        None,
+        metricsRegistry,
+        tracer)
 
       filterChain.doFilter(mockRequest, mockResponse)
